fasal-mitra/server/requirements.txt
# FastAPI Backend for FasalMitra
# Python dependencies for the server

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# HTTP & External APIs
requests>=2.31.0
httpx>=0.25.0

# AI/Chatbot
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Validation & Serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CORS & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Caching (optional)
redis>=5.0.0
aioredis>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing async endpoints

# Utilities
python-dateutil>=2.8.0
